--- When the Crow Sings/Assets/Scripts/Monobehaviours/PlayerController2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bb679c7beffc9394d9837eede24bf5b5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
