+++ When the Crow Sings/Assets/Scripts/StateMachine/StateMachine.cs
@@ -9,19 +9,8 @@ public class StateMachine
     public StateMachineState currentState;
     StateMachineState previousState;
     public Dictionary<String,StateMachineState> states = new Dictionary<String, StateMachineState>();
-    //StateMachineState[] states
-    //{
-    //    get { return GetComponents<StateMachineState>(); } // ???
-    //}
 
-    //private MonoBehaviour component;
-    public StateMachine()//MonoBehaviour monoBehaviour)
-    {
-
-        // TODO: Figure out how to get Update and FixedUpdate to be hooked up via the constructor.
-    }
-
-    // Callbacks
+    // "Callbacks"
     public void Update(float deltaTime)
     {
         currentState.Update(deltaTime);
@@ -42,10 +31,6 @@ public class StateMachine
         currentState.StateEntered();
     }
 
-    //public void RegisterState<T>(T state) where T : StateMachineState
-    //{
-    //    states[typeof(T)] = state;
-    //}
     public void RegisterState(StateMachineState state, string stateName)
     {
         states[stateName] = state;
