+++ When the Crow Sings/Assets/Scripts/StateMachine/StateMachineState.cs
@@ -1,27 +1,9 @@
-using System;
-using Unity.VisualScripting;
-using UnityEngine;
-
 public abstract class StateMachineState//<T>// where T:MonoBehaviour //: MonoBehaviour? Interface?
 {
-
-
-    // Entered Signal?
-    // Exited Signal?
-
-
     public abstract void Update(float deltaTime);
     public abstract void FixedUpdate();//(float deltaTime);
     public abstract void StateEntered();
     public abstract void StateExited();
     public abstract void OnEnable();
     public abstract void OnDisable();
-
-//    protected StateMachineState(StateMachine stateMachine, MonoBehaviour myComponent, string stateName)
-//    {
-//        this.stateMachine = stateMachine;
-
-//        //stateMachine.RegisterState<type>(this);
-//        stateMachine.RegisterState(this, stateName);
-//    }
 }
